[WebDavClient] WebDavMethodBuilder.Server requires a directory.
The problem: this would die:

	$ mono bin/Debug/webdav.exe --server https://example.com/share
	System.AggregateException:  --->
	System.AggregateException: One or more errors occured --->
	System.Xml.XmlException: Whitespace is required between PUBLIC id and SYSTEM id.  Line 1, position 55.

Rephrased, for this particular URL, an HTTP PROPFIND was behaving
like an HTTP GET, returning the _content_ at that URL (which was HTML
in this case), and _not_ the WebDAV response XML that was expected.

Oops.

Requiring that the Server base URL have a trailing '/' "fixes" the
problem, ensuring that WebDAV XML is retrieved and not the URL's
contents.