[WebDavClient, webdav] Add file download support.
Add a WebDavMethodBuilder.CreateDownloadMethodAsync() method which
will download the specified file and write to a specified Stream.

Design rationale: CreateDownloadMethodAsync() takes a Stream for the
destination instead of a filename because the PCL doesn't have a
File.OpenWrite() method (or System.IO.File!), but does have Stream, so
this will hopefully make protability easier.

Add a webdav `get SOURCE [DEST]` command.

If DEST isn't specified, writes file contents to stdout. Otherwise,
write to the specified file.