[WebDavClient] WebDavResponse methods need to check for empty values.
Consider this webdav invocation:

	webdav> ls /path/ 1 '<propfind xmlns="DAV:"><propname/></propfind>'
	Unhandled Exception:
	System.FormatException: Input string was not in the correct format
		at System.Int64.Parse (System.String s)
		at Cadenza.Net.WebDavResponse.get_ContentLength ()
		at Cadenza.Tools.WebDav.WebDavConnection.ListPath (Cadenza.Tools.WebDav.WebDavConnection state, System.String args)
		at Cadenza.Tools.WebDav.App.Main (System.String[] args)

Oops.

The problem is that the custom <propfind/> reuslts in a listing of
property names, which will include things such as <getcontentlength/>
but the elements will contain no value:

	<D:response xmlns:lp1="DAV:" xmlns:lp2="http://apache.org/dav/props/">
	  <D:href>/path/</D:href>
	  <D:propstat>
	    <D:prop>
	      <lp1:resourcetype/>
	      <lp1:creationdate/>
	      <lp1:getlastmodified/>
	      <lp1:getetag/>
	      <D:supportedlock/>
	      <D:lockdiscovery/>
	      <D:getcontenttype/>
	    </D:prop>
	    <D:status>HTTP/1.1 200 OK</D:status>
	  </D:propstat>
	</D:response>

Since the element is present, WebDavResponse.ContentLength would find
the element, but the element has no value, so we effectively try
`long.Parse("")`, which fails.

The fix is to ensure that we don't parse the empty string.